external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/construct_struct.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:7: 'constructor' :  cannot convert parameter 1 from 'int' to 'float'

implicit cast from int to float in struct initializer should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast from int to float in struct initializer should fail
test vertex shader
default fragment shaderPASS implicit cast from int to float in struct initializer should fail
PASS successfullyParsed is true

TEST COMPLETE
